requirements.txt
gradio==4.20.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
huggingface_hub==0.21.4